requirements.txt
asgiref==3.2.10
bcrypt==3.2.0
certifi==2020.6.20
cffi==1.14.3
Django==3.1.1
django-cors-headers==3.5.0
mysqlclient==2.0.1
pycparser==2.20
PyJWT==1.7.1
pytz==2020.1
selenium==3.141.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.10

